+++ src/renderutils.h
@@ -1,0 +1,23 @@
+#ifndef RENDERUTILS_H
+#define RENDERUTILS_H
+
+#include <QOpenGLFunctions_4_5_Core>
+
+/**
+ * @brief Helper class to draw a screen spaced quad
+ */
+class ScreenSpacedBuffer : protected QOpenGLFunctions_4_5_Core {
+private:
+    GLuint mVAO, mVBO;
+
+public:
+    void bind();
+    void unbind();
+
+    void draw();
+
+    ScreenSpacedBuffer();
+    ~ScreenSpacedBuffer();
+};
+
+#endif // RENDERUTILS_H
